Fix collapsible animation lag
Bootstrap has some trouble calculating the height of the collapsible
when using our `fix-shadow-overlay` class. We now avoid the use of this
class by setting padding on each element which allows the form to
display it's box shadow correctly.

Also we moved margin-top on the collapsible to a margin-bottom on the
element above because otherwise Bootstrap will have a hard time with
this one too.