Pin google chrome version to 95
Since version 96 was released to the Github CI virtual environment the
test suite was broken. Especially everything related to drag and drop
does not work anymore. It works in the browser itself. Just the tests
are failing.

After some investigation I think the root cause is the way capybara
performs html5 drag and drop simulation. There are some differences in
the behavior of the new chrome version which leads to "buggy" results
when testing drag and drop.

I'm not sure how to fix this. The problem already appeared once before
and vanished after some time so I just hope it will run at some point
with a newer chrome version...